Report JS errors to sentry.vetsgov-internal (#96)
This is the developer portal side of https://github.com/department-of-veterans-affairs/vets-contrib/issues/880

See https://github.com/department-of-veterans-affairs/devops/pull/4382 for adding the route to the rev proxy. The request looks like a POST to `https://dev-developer.va.gov/sentry/api/15/store/?sentry_key=7d7f3fde3b744e7099dc5e8cc65b01f7&sentry_version=7`

It's likely that there will be more customization to do to make these reports as useful as possible, but it's hard to do that before we deploy this. Open to any suggestions now if you have some sentry experience, otherwise we can iterate on this later.