Update formation and formation react to latest versions (#167)
* Update formation and formation react to latest versions

* Reenable accessibility testing for /apply page

Need to add an id to the header section to pass accessibility. Cannot
add an id directly via the markdown-component-loader. Wrapped the
stateless component in an extra div

* Move apply header to an actual component

Apply header is no longer a markdown component. There was a lot of HTML
tags in the markdown file and we were going to need to add another (# ->
<h1>). Switched the component to not be a markdown component since it
mostly wasn't anyways.