Extract sidenav component (#174)
* Begin extracting sidenav components

* Fix sticky nav

* Prefer const

* Fix import path

* Minor sidenav component cleanup

* Add comment explaining stickyfill ref

* Remove unnecessary div

* Cleanup/format exploredocs

* Make styles less specific where possible

* Use sidenav on support page

* Fix selector in e2e test

* Update vr screenshots

* Fix logic error in hashmatch

* Move hashMatch function to public instance method

* Add unit test of hashmatch function

* Add one more test case

* Give function more accurate name

* Fall back to native matching function

When `exact == false` and `to` is not a matching hash, use the
react-router isActive implementation, which does partial matching using
path-to-regexp

* Upgrade enzyme and types

* Test SideNavEntry component with router

Since we're depending on the NavLink isActive as a fallback, we need a
real match object to be passed into our hashIsActive function. Static
rendering a router lets us do that.

* Only import variables

* Fix sidenav bugs and improve tests

Adds a few new tests for some missed logical bugs. Also allows both
the object and string form of the `to` prop

* Add explanatory comment

* Add additional test case