Update SupportContactUs.tsx
Additional edit--the link to the Apple Health support page should go up here, because if we're looking to remove the GitHub option, the only thing under the Contact Us form should be an indication that red asterisked fields are required. Let me know if I'm making sense or if you want to talk through this a bit more.