Api 2496: Hook migration for ApiPage, VeteransCrisisLine (#442)
* Use React Hooks for ApiPage

There were some non-null assertions in here that eslint does not like. I
removed them but have questions on how we should handle these cases, if
they are somehow null - or if we have non-null assertions, I assume we
have a good reason. So maybe we need to do some refactoring in order to
solve those issues.

* Use ReactHooks for VeteransCrisisLine

To try and clean up the component I made a custom hook for controlling
the modal. I didn't see any other modals in the code, or at least not
obvious modals, to have a reason to pull the custom Effect out into a
different file. Maybe we could create an effects/ folder in our src code
to store any custom effects we create for easy reuse?

* Extra logic into custom hook - ModalController

I researched a bit to see how we can make our code a little cleaner as
we move things over to hooks, and the pattern I found was to make custom
effects as needed. This helps extract function to clean up our files.

I named the custom hook ModalController in this case because it's
essentially a controller for a modal, and couldn't really thinkg of a
better name. - File is called ModalController but exports
useModalController.

I did include a few eslint fixes in this commit as well.

* Add type to svg files when imported

I was running into an eslint issue where imported svg files had the any
type. I resolved this by exporting the default value (path) as a string.

* Add PropTypes to react hooks

Note that on ApiPage, the propTypes are set to an empty object, but it
actually takes some Routing properties. This is because we want to
upgrade ApiPage to use hooks but at present we're on version 4 of
react-router-dom. In order to use the hooks, which will remove the need
for us to have any props passed to ApiPage, we need to upgrade to
version 5 of react-router-dom.

* Handle undefined values with content

In order to resolve eslint issue I removed non-null assertions [The (!)
operator]. This commit adds better backup content if the values are
undefined or null.