Api 2500: Hooks migration CategoryReleaseNotes, ServersContainer, OperationTag (#437)
* Remove unused swagger components

We didn't uses the Servers or ServersContainer component so they are now
removed. I did, however leave the styling in StylesOverride.css for the
scheme-container that these components existed inside. Without that
override the scheme-component div element created an unnatural looking
blank space on the screen.

OperationTag was left unchanged because it's a swagger plugin, and
swaggerUI has issues with react 16 at this time.