Api 2546: Hook migration for banner (#447)
* Convert Banner to use react hooks

* Add mocks for images for unit tests

With typescript and requiring properties to have values, I found images
were being imported as undefined within unit tests. This fixes unit
tests. I still need to look into other tests. I'm particularly concerned
there might be an issue with visual regression tests, but we'll find
out!

* Add alt text property to icon

The icons in the GuidanceBox component all had the same alt text. This
adds an optional icon_alt property to set the alt text on the given
icon. It's optional so that if the icon is purely decorative, we can
leave the alt text off.