API-2943: Add ModalController test (#487)
* Add ModalController test

I tried to use the renderHook function from the testing library but ran
into issues with eslint claiming it was using an any type, even after I
specified the types on everyting. For the moment I went with the pattern
used on the Previous hook we made. For the most part, if we switch to
the renderHook function at a later date, the test shouldn't really
change besides how we are rendering

* Add test coverage for custom hooks

The custom hooks now have better test coverage.

I added the react-hooks testing library so we can use the renderHook
function for testing our custom hooks. That said, we aren't using this
in ModalController because it requires access to the DOM for firing
'typing' events. I couldn't find a way to access them (screen doesn't
work because react-hooks uses the react-test-renderer for rendering and
works differently.

* Add new line at end of file

Co-authored-by: Tyler Snell <tylersnell@me.com>