API-3029: Set colors so contrast is at least 4.5:1 (#552)
* Set colors so contrast is at least 4.5:1

In order to do this, I overrided some css classes in the theme we are
using (atom-one-dark-reasonable). I created a separate scss file for the
overrides so they can be tracked separately at a later date, and if the
original theme is updated, it will be easier to track when we can remove
those overrides.

* Explain why contrast checks are disabled on auth

I updated the comment explaining why we have color contrast checks
turned off on the authorization page, in the accessibility automated
tests.

Once a new version of jest-axe comes out with axe-core 4.1.0 as a
downstream dependency (and we upgrade), we should be able to turn these
checks back on. Axe-core 4.1.0 documents that it fixed this issue with
color contrast checks.