Api 3786: Use semantic tags on main page (#562)
* Use semantic paragraphs in banner guidance boxes

Using semantic html paragraphs rather than divs increases the usability
of the site for those using screen readers.

* Use p tags for paragraph content

This commit includes keeping everything visually consistent. There will
likely be future styling updates that slightly modify the layout of the
page.

* Remove break tag from main page

I removed the 'no margin' css class from the paragraph tags on the main
page. The main page now has proper margins on the paragraphs and so
there is no need for a break tag.

* Update tests for minimal visuals on semantic html

I had difficulty making succinct title for this. Due to the semantic
html changes the styling was changed very slightly on the main page.
This is due to the paragraph tags having margins. The visual tests are
updated now with regard to those changes