API-3758: Add test for MarkdownPage (#570)
* Add test for MarkdownPage

* Use queryBy for testing existence of elements

The testing library 'getBy' methods throw errors when an element does
not exist. This seems appropriate in scenarios where we want to get an
element and check its properties. However, when checking for existence
of an element, we want to be explicit and have an 'expect' statement. We
can accomplish this by using 'queryBy' which returns null and then
expecting the response to not be null. This lets us check the existence
without throwing errors.

* Remove unneeded import of extend-expect

Co-authored-by: Tyler Snell <tylersnell@me.com>