Resolve missing Link component (#587)
The ApiDocumentation test was having issues that the React 'Link' tag
was not defined. This was caused by a mock of the react router dom. It
includes 'useHistory' and 'useLocation' but not 'Link'. To rememdy this,
we can use jest.requireActual and a spread operator to effectively only
mock those two methods of the node library and use the native package
versions of Link, useParams, and any other values of react-router-dom.