API 2991: Respect tier 2 slugs - display Not Found when category does not have api (#574)
* Display Not Found when category does not have api

When a user searches for an api under a category that does not have an
api of the given name in the url, we now display a 404 not found page.

For example, the health category does not include a claims api. Before
when a user would go to /explore/health/docs/claims, the claims api
under the benefits category would be displayed (confusingly also
identifying itself as a health category api). Now when a user searches,
they get a Not Found page.

* Simplify Not Found return logic

I collapsed the two separate return statements to one to help keep this
code with two separate returns a little more readable.

* Refactor code - comment/explain unreachable code

In the deactivation message of the ApiPage component, we have two
methods used to determine if an api is deprecated or deactivated. Both
of these methods return false if the api's deactivation info is
undefined. In this scenario, DecativationMessage was returning null.
Afterward in the code, the compiler does not see that api
deactivationInfo must be defined to get past that if statement that
returns null. This caused us to add additional if logic, and create
arrow functions for rendering in cases that are theoretically
unreachable.

While writing the test for this component, I was unable to reach 95%
coverage due to this. This refactor should allow me to reach 95%
coverage, and really simplify our logic.

I put a large comment explaining why the new theoretically unreachable
code exists and that, if for some reason, my logic on this isn't sound,
that that line might be a good place to start searching if
DeactivationMessages were to start disappearing from pages that should
have them

* Add ApiPage test

This is the first commit for this test with a 95% coverage.

A couple notes:
- I mocked the ExplorePage, since it's an .mdx file, and it didn't seem
to be working properly in the React/Jest ecosystem.
- I mocked the ApiDocumentation component, since it was having display
issues due to not having the entire reach store configured in this test.
- I mocked the react router dom to easily change the parameters I wanted
it to have. I may removed this mock, if possible, in a later commit.
- I add an openAPI server. This may be unneeded. I'm looking into
removing this.
- I attempted to make the tests verbose. This lead to 4 indents, which
our linter does not like. We could refactor these tests to resolve the
issue.
- I used similar 'spyOn' functionality that the already existing
ApiDocumentation test uses. With this I was able to mock the existing
categories and apis without needing to mock the actual apiDefs files.

* Cast types as little as possible

* Refactor to remove need for disabling eslint rules

* Remove unused mocked web server

* Use more explicit functions in tests

I was using expect().toBeDefined(), but null is considered defined. In
this test, I think we want to check that the element is not null. I
swiched the test to now use expect().not.toBeNull();

* Test deactivation message. Remove router dom mock

The code now has a basic test that the deactivation message is not shown
for an API wih deactivation info but has not met it's deactivation date
yet.

This also removes an unneeded mock for the react-router-dom and some
code refactoring as a result of that.

* Get 95% branching coverage on Api Page test

I added some more tests to reach 95% coverage. I think we can likely
refactor this test to make it cleaner.

One thing I'd like to do is to maybe make a category builder so that we
can easily create mock categories as needed in our tests. I also think
it may be beneficial, this being a unit test, to restructure this to
check one branch at a time rather than the way it is currently written,
with each scenario testing various pieces of output. - ie. I want to
have the unit test contain various unit tests rather than having it
contain various small scale e2e/integration tests.

* Make comment more clear

* Clarify some tests

There was one test that had the wrong description.

I added convenience variables for the fake categories, so there
weren't random index values through the test. Note that if
fakeCategories are updated, then this test will need updated
appropriately. For instance, if the indexes change or if an
existing category/api is modified in a way that the test cases break

I removed the semi-global BeforeEach and have each inner describe
setting up its own mocks. Because of how long this test file is, this
helps make it more explicit within each test the scenario and that the
mocked values match that description of the test case.

Co-authored-by: Tyler Snell <tylersnell@me.com>