Upgrade react-dev-utils to 11.0.1 (#609)
* Upgrade react-dev-utils to 11.0.1

This upgrades react-dev-utils to 11.0.1.

Note that e2e and visual regression tests are not running properly with
this change. A later commit will be made to address this.

API-4498

* Downgrade react-dev-utils to 9.1.0

With a higher version of react-dev-utils the e2e and visual regression
tests were failing. With version 9.1.0 they pass once again.

For now we're downgrading and will upgrade again at a later date. The
strategy will likely be to step one version at a time to determine if
there's something in particular that is causing the issues.

API-4498

* Upgrade react-dev-util to 11.0.1 and stabilize

Upgraded react-dev-utils to 11.0.1, the latest.
Added two options to the webpackDevServer.config.js to make the server
stay more stable. After researching why the server would get errors, I
found the dev server was upgraded to use native sockets rather than
sockjs-node sockets. In order to get things stable, we needed to set
transportMode to 'ws'. I also added an option 'injectClient' and set it
to false based on information I found in the same place. It should
prevent the server from attempting to inject a WS client when we already
have the dev server injected.

API-4498

* Clarifying comment that we have no proxy

Added a clarifying comment about the prepareProxy function and specified
that it comes from the original create-react-app configuration.

API-4498

Co-authored-by: Tyler Snell <tylersnell@me.com>