API-5230: upgrade markdown it dependencies (#637)
* All dependencies upgraded to the latest version the app supports.

* Added back file SitemapBuilderWebpackPlugin.js

* Fixed eslint errors in test files with the await command.

* Use webpack compiler hook api

Removed deprecated code in favor of the new hook api

API-4289

* Updated to latest 22 dependencies.

* Upgraded msw fixed type errors.

* Fix husky pre commit hooks

* Added husky 5 configuration.

* Removed husky v4 pre commit hook config.

* Upgrade to latest webpack 4

We are unable to upgrade to webpack 5 at the current time because we use
react-dev-utils and it has a bug where it parses messages improperly.
For now the best we can do is upgrade to the latest version of webpack
4.

API-4494

* Upgrade webpack cli and dev server to latest

Both the webpack cli and webpack dev server have been updated to their
latest versions. They appear to work fine but I need to test in another
environment. The dev build in my local seems to have an issue loading
the api specifications, but I believe this is due to the way I am
running it and not a bad package. This will be tested out and specified
in the PR

API-4493

* Reword comment

* Upgrated packages.

* Upgraded ts loader and url loader

* Changed tsConfig field to tsconfig field to solve deprecated warnings.

* Solved merge conflicts of package.json with master

* Upgraded webpack manifest plugin and updated named import.

* Updated html-loader

* Updated file-loader dependency.

* Migrated deprecated library @babel/polyfill to core-js and regenerator-runtime per babel docs suggestion.

* Added changes based on review.

* Updated packages that just got new versions.

* Upgraded markdown-it related dependencies

* Added fix for npm-force-resolution package.

Co-authored-by: Anthony Kopczyk <anthony.kopczyk@oddball.io>