API-3904: Dependabot Setup (#631)
* Initial commit for dependabot setup

* Remove target branch from dependabot config

I had target branch in the dependabot config to try and test it out, but
I'm not sure how easy it will be to test. I think at this point it might
be best to merge this, and then on the first of next month see if the
dependabot updates come up correctly.

* Set custom labels on dependabot PRs

This sets label for npm dependabot PRs to be npm, dependencies, and
dependabot. Note these labels only apply to the npm package ecosystem in
the configuration. If we add other package ecosystems we'll need to
specify those labels separately.

* Ignore packages that we can not currently upgrade

This adds configuration to dependabot telling it to ignore dependenciese
we can't currently upgrade. It also adds descriptions of why we are
unable to upgrade.

* Use single quotes for non-strings

Yaml tries to assume the types of the values it gets. In this scenario
starting with an '@' does not get read as a string. I added single
quotes around @@department-of-veteran-affairs/formation-react to resolve
this formatting problem.

* Check weekly for dependabot updates

We're going to be updating packages weekly or bi-weekly, so we want
dependabot configured for weekly checks rather than monthly.