API-5574: upgrade react to 17 (#712)
* Upgrade react to version 17

This upgrades react to version 17

Note that this changes some code in the SwaggerDocs components to prevent
incorrect versions from being set. This was caused because of React 17s
modification to how it treats cleanup useEffect functions

API-5574

* Fix bug with benefits side nav links

This is one commit to fix the benefits side nav links not updating. This has
linting errors that will need to be solved

API-5574

* Use react callback for setting doc metadata

This moves the setMetadataAndDocUrl function for SwaggerDocs into a
'useCallback'. This appears to resolve the issue where the docs do not update
between the benefits intake and benefits claims apis. It also appears to allow
the other api pages to function properly

API-5574

* Use accurate names

The name for the method 'getInitialVersion' was accurate but with the latest
changes to the code base it is now misleading. This updates it to fit with the
newest code

API-5574

* Organize code in SwaggerDocs

This organizes some of the code in the SwaggerDocs component into logical
sections. For instance, the code related to the api version is moved closer to
the 'useEffect' call for handling version updates

API-5574

Co-authored-by: Alexander Valencia <alex.valencia@adhocteam.us>