Fix spacing with error elements + nested conditional fields in prod access form (#842)
* Space form fields correctly on prod + sandbox access forms
- Ensure that fields with .usa-input-error are statically positioned so they don't move left when there is an error
- Move nested conditional fields 32px to the right to make clear they are under the question that triggers their appearance

* More spacing fixes

* Remove unnecessary error boundaries from text inputs (input[type="text"] and textarea) that are children of fieldsets with the usa-input-error class

The way this usually happened, at least in the prod access form, is that there would be one radio button group that triggered one or more nested conditional fields within the same fieldset. If the user validated the form with no option selected (usually Yes or No), then the whole fieldset would have an error. Then, if the Yes option was selected and the conditional fields became visible, they would still have the red error border around the input because of the usa-input-error on the fieldset. To solve this problem, we explicitly set the border of the input to be 1pxsolid gray any time that it does not have the usa-input-error class itself.